+++ R/plot_ptables.R
@@ -1,0 +1,604 @@
+#' @title Plot Functions for a List of Parameter Tables
+#'
+#' @description Various plot functions
+#' for the output of [eq_models()] and
+#' similar functions.
+#'
+#' @details
+#' The functions provide different ways
+#' to visualize the models. Basic knowledge
+#' of [semPlot::semPaths()] from the
+#' package `semPlot` is required.
+#'
+#' @seealso See [eq_models()] on the
+#' type of output supported.
+#'
+#' @examples
+#' library(lavaan)
+#'
+#' # Model 1
+#'
+#' mod1 <-
+#' "
+#' fx =~ x1 + x2 + x3
+#' fm =~ m1 + m2 + m3
+#' fy =~ y1 + y2 + y3
+#' fm ~ fx
+#' fy ~ fm + fx
+#' "
+#' fit1 <- sem(
+#'           model = mod1,
+#'           data = data_test_3_factor_3_item
+#'         )
+#' fit1_1_more <- drop_k(fit1)
+#' fit1_1_more_1_less <- lapply(
+#'   fit1_1_more,
+#'   add_k
+#' )
+#'
+#' partables1 <- combine_partables(fit1_1_more_1_less)
+#' eq_out_1 <- eq_models(
+#'           partables1,
+#'           original_model = fit1,
+#'           parallel = FALSE
+#'         )
+#'
+#' eq_out_1 <- eq_models(
+#'           partables1,
+#'           original_model = fit1,
+#'           parallel = FALSE
+#'         )
+#'
+#' layout_i <- matrix(c(  NA, "fm",  NA,
+#'                      "fx",   NA, "fy"),
+#'                    ncol = 3,
+#'                    nrow = 2,
+#'                    byrow = TRUE)
+#' layout_i
+#' p <- partables_plots(
+#'   eq_out_1,
+#'   original_model = fit1,
+#'   layout = layout_i,
+#'   label.cex = 1.8,
+#'   sizeMan = 11,
+#'   edge.width = 5,
+#'   asize = 5,
+#'   structural = TRUE,
+#'   par_diff_settings = list(
+#'             color = "blue",
+#'             width = 10
+#'           )
+#' )
+#' plot(p)
+#'
+#'
+#' @name plot_partables
+NULL
+
+
+#' @return
+#' The function [partables_plots()] returns
+#' a list of `qgraph` objects generated
+#' from [semPlot::semPaths()], of the
+#' class `partables_plots`.
+#'
+#' @param partables It should be a list of
+#' parameter tables, such as the output
+#' of [eq_models()] or [eq_df_models()].
+#'
+#' @param original_model The model to
+#' which models in `partables` will be
+#' compared to. If `NULL`, then the
+#' plots will be generated without checking
+#' for differences between a model and
+#' `original_model`.
+#'
+#' @param ... Optional arguments to be
+#' passed to [semPlot::semPaths()]. They
+#' will be applied to all plots.
+#'
+#' @param par_diff_settings A named list of
+#' settings to configure the change of
+#' a parameter from `original_model`
+#' to a model in `partables`. For now,
+#' two settings are supported: `color`
+#' for the color and `width` for the
+#' width of an edge (arrow).
+#'
+#' @param fix_pars_fixed_zero If `TRUE`,
+#' parameters fixed to zero will be
+#' modified based on `par_fixed_zero_settings`.
+#'
+#' @param par_fixed_zero_settings A named
+#' settings to configure edges (arrows)
+#' fixed to zero. For now,
+#' two settings are supported: `color`
+#' for the color and `width` for the
+#' width of an edge (arrow). Setting the
+#' width to zer0, the default, effectively
+#' hide an arrow.
+#'
+#' @param exclude_original_model If `TRUE`
+#' and `original_model` is set, the output
+#' will not include the plot of the original
+#' model, though this plot will be stored
+#' in the attribute `"original_model"`,
+#' as a one-element list.
+#'
+#' @param auto_node_color If `TRUE`,
+#' and the number of nodes is less 12 or
+#' less, they will be automatically
+#' set to different colors. Ignored if
+#' `color` of nodes is in `...`.
+#'
+#' @rdname plot_partables
+#' @export
+partables_plots <- function(
+  partables,
+  ...,
+  original_model = NULL,
+  auto_node_color = TRUE,
+  par_diff_settings = list(
+              color = "blue",
+              width = 2
+            ),
+  fix_pars_fixed_zero = TRUE,
+  par_fixed_zero_settings = list(
+              color = "white",
+              width = 0
+            ),
+  exclude_original_model = TRUE
+) {
+  if (!is_partables(partables)) {
+    stop("partables is not of a supported type.")
+  }
+
+  if (!is.null(original_model)) {
+    if (!is_partable(original_model)) {
+      if (!inherits(original_model, "lavaan")) {
+        stop("original_model is not of a supported type.")
+      }
+      original_model <- lavaan::parameterTable(original_model)
+    }
+
+    # ==== Generate Model Differences ====
+
+    f0 <- function(
+      pt_i,
+      original_model
+    ) {
+      pt_diff_i <- model_diff(
+        original_model,
+        pt_i
+      )
+      diff_i <- pt_diff_i$pt_i
+      if (any(diff_i$free > 0)) {
+        new_par <- lavaan::lav_partable_labels(diff_i)
+      } else {
+        new_par <- NULL
+      }
+      new_par
+    }
+    pt_pars_diff <- lapply(
+      partables,
+      FUN = f0,
+      original_model = original_model
+    )
+  } else {
+    pt_pars_diff <- vector("list", length(partables))
+  }
+
+  # ==== Generate the Plots ====
+
+  out <- mapply(
+    FUN = partables_plots_internal,
+    pt = partables,
+    pars_diff = pt_pars_diff,
+    MoreArgs = c(
+      list(...),
+      list(
+        auto_node_color = auto_node_color,
+        fix_pars_fixed_zero = fix_pars_fixed_zero,
+        par_fixed_zero_settings = par_fixed_zero_settings,
+        par_diff_settings = par_diff_settings
+      )
+    ),
+    SIMPLIFY = FALSE,
+    USE.NAMES = TRUE
+  )
+
+  if (exclude_original_model &&
+      !is.null(original_model)) {
+    tmp <- partables %pt_in% as_eq_partables(original_model)
+    if (any(tmp)) {
+      out_org <- out[tmp]
+      out <- out[!tmp]
+      attr(out, "original_model") <- out_org
+    }
+  }
+
+  class(out) <- c("partables_plots", class(out))
+  out
+}
+
+#' @return The `plot` method of the output
+#' of [partables_plots()] return `x`
+#' invisibly. Called for its side effect.
+#'
+#' @param x The output of [partables_plots()],
+#' a `partables_plots` object.
+#'
+#' @param ... Optional arguments to be
+#' passed to [semPlot::semPaths()].
+#'
+#' @param title_mode What will be used
+#' as the title. If `"digest"`, then
+#' the digest value of a model will be
+#' used as its title. If `"name"`, then
+#' its name in `x` will be used. If
+#' `"none"`, then no title will be drawn
+#' with  the plot.
+#'
+#' @param title_adj Adjust the position
+#' of the title. Increase this value
+#' if the plot is too close to the title,
+#' and decrease this value if the space
+#' between the plot and its title is
+#' too large.
+#'
+#' @param title_args A named list of
+#' arguments to be passed to [title()]
+#' when printing the title.
+#'
+#' @param ncol,nrow The number of columns
+#' and rows when drawing the models. Used
+#' by `mfrow` in [par()].
+#'
+#' @param scale_plots How the models will
+#' be scaled. If `"auto"`, then the models
+#' will be scaled based on `ncol` and `nrow`,
+#' using `scale` as a reference. If
+#' `"always"`, then `scale` will always
+#' be used to scale the models, even if
+#' they are drawn one by one.
+#'
+#' @param scale How the model will
+#' be furthered scaled when drawn. If this
+#' value is greater than one, then
+#' elements in `elements_to_scale` will
+#' be increased by this ratio. If this
+#' value is less than one, then elements
+#' in `elements_to_scale` will be decreased
+#' by this ratio.
+#'
+#' @param elements_to_scale Elements
+#' in the `qgraph` object that will be
+#' scaled based on the values of
+#' `scale_plots` and `scale`.
+#'
+#' @param original_model_mode How original
+#' model, if present will be handled when
+#' drawing the models. If `"exclude"`,
+#' then the original model will not be
+#' drawn. If `"include"`, the original
+#' model will be drawn as the first model,
+#' along with other models. If `"side_by_side"`,
+#' then number of columns is always two
+#' and the number of rows is always one.
+#' In each plot, the original model will
+#' be drawn on the left and the other
+#' model will be drawn on the right.
+#'
+#' @rdname plot_partables
+#' @export
+plot.partables_plots <- function(
+  x,
+  ...,
+  title_mode = c("digest", "name", "none"),
+  title_adj = 1.4,
+  title_args = list(),
+  ncol = 1,
+  nrow = 1,
+  scale_plots = c("auto", "always"),
+  scale = NULL,
+  elements_to_scale = c(
+    "Nodes:width",
+    "Nodes:height",
+    "Edges:width",
+    "Edges:asize"
+  ),
+  original_model_mode = c("exclude", "include", "side_by_side")
+) {
+  original_model_mode <- match.arg(original_model_mode)
+  scale_plots <- match.arg(scale_plots)
+  title_mode <- match.arg(title_mode)
+
+  if (original_model_mode == "side_by_side") {
+    nrow <- 1
+    ncol <- 2
+  }
+  parold <- par(mfrow = c(nrow, ncol), no.readonly = TRUE)
+
+  original_model_list <- attr(x, "original_model")
+  has_original_model <- !is.null(original_model_list)
+
+  original_title_suffix <- "(Original)"
+  if (has_original_model) {
+    original_digest <- get_digest(original_model_list[[1]])
+  } else {
+    original_digest <- ""
+  }
+
+  if (has_original_model) {
+    x <- c(original_model_list,
+           x)
+  }
+
+  # ==== auto_scale ====
+  if (scale_plots == "auto") {
+    if (((ncol != 1) || (nrow != 1))) {
+      scale_i <- scale %||% c(x = ncol / 2, y = nrow / 2)
+      if (length(scale_i) == 1) {
+        scale_i <- c(x = scale_i, y = scale_i)
+      }
+    } else {
+      scale_i <- c(x = 1, y = 1)
+    }
+  }
+  if (scale_plots == "always") {
+    scale_i <- scale %||% c(x = 1, y = 1)
+    if (length(scale_i) == 1) {
+      scale_i <- c(x = scale_i, y = scale_i)
+    }
+  }
+  if (any(scale_i != 1)) {
+    x <- lapply(
+      x,
+      FUN = scale_plot,
+      scale_x = scale_i["x"],
+      scale_y = scale_i["y"],
+      elements = elements_to_scale
+    )
+  }
+
+  if (has_original_model) {
+    if (original_model_mode == "exclude") {
+      x <- x[-1]
+    }
+    if (original_model_mode == "include") {
+      # A placeholder
+    }
+    if (original_model_mode == "side_by_side") {
+      k0 <- seq_along(x)
+      k1 <- sapply(
+              k0[-1],
+              \(x) c(1, x)
+            )
+      x <- x[k1]
+    }
+  }
+
+  # ==== Plot the model ====
+
+  for (i in seq_along(x)) {
+    xx <- x[[i]]
+    xx_name <- names(x)[i]
+    if (title_mode != "none") {
+      xx$plotOptions$mar[3] <- xx$plotOptions$mar[3] * title_adj
+    }
+    plot(xx,
+         ...)
+    xx_digest <- attr(xx, "digest")
+    tmp <- switch(
+      title_mode,
+      digest = xx_digest,
+      name = xx_name,
+      none = NULL
+    )
+
+    if (xx_digest == original_digest) {
+      tmp <- paste(tmp, original_title_suffix)
+    }
+
+    if (!is.null(tmp)) {
+      do.call(
+        match.fun("title"),
+        c(list(main = tmp),
+          title_args)
+      )
+    }
+  }
+  par(parold)
+  invisible(x)
+}
+
+#' @noRd
+partables_plots_internal <- function(
+  pt,
+  ...,
+  fix_cov = FALSE,
+  cov_settings = list(
+              color = "red"
+            ),
+  auto_node_color = TRUE,
+  pars_diff = NULL,
+  par_diff_settings = list(
+              color = "blue",
+              width = 2
+            ),
+  fix_pars_fixed_zero = TRUE,
+  par_fixed_zero_settings = list(
+              color = "white",
+              width = 0
+            )
+) {
+  # Internal function (for now)
+  # Generate the plot for a model
+
+  pt$lavlabel <- lavaan::lav_partable_labels(pt)
+
+  ddd <- list(...)
+  # Can be overriden
+  semPaths_default <- list(
+    residuals = FALSE
+  )
+  ddd1 <- utils::modifyList(
+    semPaths_default,
+    ddd
+  )
+  # Cannot be overriden
+  ddd1 <- utils::modifyList(
+    ddd1,
+    list(
+      object = pt,
+      DoNotPlot = TRUE
+    )
+  )
+
+  p_fit <- do.call(
+    semPlot::semPaths,
+    ddd1
+  )
+
+  if (is.null(ddd$color) &&
+      auto_node_color) {
+
+    # ==== Set the colors for nodes ====
+
+    k_nodes <- length(p_fit$graphAttributes$Nodes$color)
+    node_names <- p_fit$graphAttributes$Nodes$names
+
+    if (k_nodes <= 12) {
+      # TODO:
+      # - Try paletteer_d("ggsci::default_ucscgb")
+      color_nodes <- RColorBrewer::brewer.pal(k_nodes, "Set3")
+      # Ensure the assignment of colors are the same
+      # across models.
+      names(color_nodes) <- sort(node_names)
+      color_nodes <- color_nodes[node_names]
+      ddd1 <- utils::modifyList(
+        ddd1,
+        list(color = color_nodes)
+      )
+      p_fit <- do.call(
+        semPlot::semPaths,
+        ddd1
+      )
+    }
+  }
+
+  if (fix_cov) {
+
+    # ==== Covariance edges ====
+
+    p_fit_vars <- names(p_fit$graphAttributes$Nodes$labels)
+    i <- (pt$lhs %in% p_fit_vars) &
+         (pt$op == "~~") &
+         (pt$lhs != pt$rhs)
+    if (any(i) && fix_cov) {
+      covs <- pt$lavlabel[i]
+      cov_colors <- rep(cov_settings$color, length(covs))
+      names(cov_colors) <- covs
+      p_fit <- semptools::set_edge_color(
+                  p_fit,
+                  color_list = cov_colors
+                )
+    }
+  }
+
+  if (fix_pars_fixed_zero) {
+
+    # ==== Paths fixed to zero ====
+
+    i_fixed_zero <- (pt$free == 0) &
+                    (pt$op %in% c("~", "~~", "=~")) &
+                    (pt$start == 0)
+    if (any(i_fixed_zero)) {
+      pars_fixed_zero <- pt$lavlabel[i_fixed_zero]
+      tmp1 <- par_fixed_zero_settings$color
+      tmp2 <- par_fixed_zero_settings$width
+      for (xx in pars_fixed_zero) {
+        names(tmp1) <- xx
+        names(tmp2) <- xx
+        p_fit <- tryCatch(
+            semptools::set_edge_color(
+              p_fit,
+              color_list = tmp1
+            ),
+            error = function(e) p_fit
+          )
+        p_fit <- tryCatch(
+            semptools::set_edge_attribute(
+              p_fit,
+              values = tmp2,
+              attribute_name = "width"
+            ),
+            error = function(e) p_fit
+          )
+      }
+    }
+
+  }
+
+  # ==== Parameter Differences ====
+
+  if (!is.null(pars_diff)) {
+    tmp1 <- par_diff_settings$color
+    tmp2 <- par_diff_settings$width
+    for (xx in pars_diff) {
+      names(tmp1) <- xx
+      names(tmp2) <- xx
+      p_fit <- tryCatch(
+        semptools::set_edge_attribute(
+          p_fit,
+          values = tmp1,
+          attribute_name = "color"
+        ),
+        error = function(e) p_fit)
+      p_fit <- tryCatch(
+        semptools::set_edge_attribute(
+          p_fit,
+          values = tmp2,
+          attribute_name = "width"
+        ),
+        error = function(e) p_fit)
+    }
+  }
+
+  # ==== Additional Info ====
+
+  attr(p_fit, "digest") <- get_digest(pt)
+  attr(p_fit, "gen_models_name") <- attr(pt, "gen_models_name")
+
+  p_fit
+}
+
+#' @noRd
+scale_plot <- function(
+  object,
+  scale_x = 1,
+  scale_y = 1,
+  elements = c(
+#    "Nodes:label.cex",
+    "Nodes:width",
+    "Nodes:height",
+#    "Edges:label.cex",
+    "Edges:width",
+    "Edges:asize"
+  )
+) {
+  out <- object
+  for (xx in elements) {
+    xx1 <- strsplit(xx, ":")[[1]]
+    scale1 <- switch(
+      xx1[2],
+      width = max(scale_x, scale_y),
+      height = max(scale_y, scale_x),
+#      label.cex = min(scale_x, scale_y),
+      max(scale_x, scale_y)
+    )
+    out$graphAttributes[[xx1[1]]][[xx1[2]]] <-
+      out$graphAttributes[[xx1[1]]][[xx1[2]]] * scale1
+  }
+  out
+}